Fix: Centralize BasePrompt and resolve import issues across notebooks
This commit addresses issue #21 by fixing the BasePrompt import issues
and improving the notebook infrastructure.

Changes:
- Centralized BasePrompt: All notebooks now import from
  agentic_platform.core.models.prompt_models instead of local utils
- Added missing methods to central BasePrompt:
  * to_bedrock_messages() for message formatting
  * to_bedrock_system() for system prompt formatting
- Removed deprecated labs/module2/notebooks/utils/base_prompt.py
- Fixed import order in agentcore_tool_browser notebook
- Added nova-act>=0.1.0 dependency to pyproject.toml
- Updated module4 README with API key setup instructions for
  Tavily and Nova Act
- Added .gitattributes for nbstripout to handle notebook metadata
- Fixed ActResult handling in browser tool for proper result extraction

Notebooks updated:
- labs/module2/notebooks/3_routing.ipynb
- labs/module2/notebooks/4_parallelization.ipynb
- labs/module2/notebooks/5_orchestrator.ipynb
- labs/module2/notebooks/6_evaluator_optimizer.ipynb
- labs/module4/notebooks/2_multi-agent-delegation.ipynb
- labs/module4/notebooks/5_agentcore_tool_browser.ipynb
