+++ .gitattributes
@@ -1,0 +1,2 @@
+*.ipynb filter=nbstripout
+*.ipynb diff=ipynb
