+++ pyproject.toml
@@ -47,6 +47,7 @@ dependencies = [
     "psycopg2-binary>=2.9.10",
     "langchain-litellm>=0.2.2",
     "bedrock-agentcore>=1.0.3",
+    "nova-act>=0.1.0",
 ]
 
 [tool.uv.workspace]
